Bug fix: initialize measured snow height with nodata, such that it propagates through SNOWPACK as nodata when no measured HS is provided in the forcing data.
This is an attempt to fix a bug where the measured snow height is set to initial snow height from the *sno file, without further updates when measured snow height is not available in the forcing data. For example: when SNOW_EROSION is then set to TRUE, the snow drift routine will make a wrongful comparison of measured vs computed snow heights.