Conserving ILWR when ILWR is provided in the input data.
ILWR is translated in emissivity EA. However, EA is standard capped at 1.0, which introduces unwanted discrepancies between ILWR provided as input, and ILWR used in the computations. This commit fixes that situation: EA is only capped at 1.0, when the source of ILWR is a parameterization or approximation. If ILWR is provided as input data, it is assumed to be of such good quality, that it can be applied directly. See also point 3 in issue #868.